requirements.txt
feedparser
python-dotenv
requests
beautifulsoup4
google-generativeai
pytrends
openai
anthropic
playwright
Pillow

